Moved local_values and statistics to other modules; renamed gradient to gradient_of_expectation
