+++ Sources/Stats/obs_manager.hpp
@@ -109,7 +109,7 @@ class ObsManager {
   }
 };
 
-void to_json(json &j, const ObsManager &om) {
+inline void to_json(json &j, const ObsManager &om) {
   auto names = om.Names();
   j = json();
   for (auto name : names) {
